+++ Sources/SwiftyHYGDB/RadialStar.swift
@@ -30,8 +30,8 @@ extension RadialStar {
             var right_ascension = Float(fields[7]),
             var declination = Float(fields[8]),
             let dist = Double(fields[9]),
-            let mag = Double(fields[13]),
-            let absmag = Double(fields[14])
+            let mag = Float(fields[13]),
+            let absmag = Float(fields[14])
             else {
                 print("Invalid Row: \(row), \n fields: \(fields)")
                 return nil
@@ -133,7 +133,7 @@ extension RadialStar: Equatable {}
 extension RadialStar: CustomDebugStringConvertible {
     public var debugDescription: String {
         let distanceString = starData?.value.distance ?? Double.infinity
-        let magString = starData?.value.mag ?? Double.infinity
+        let magString = starData?.value.mag ?? Float.infinity
         return "🌠: ".appending(starData?.value.properName ?? "N.A.")
             .appending(", Hd(\(starData?.value.hd_id ?? -1)) + HR(\(starData?.value.hr_id ?? -1))")
             .appending("Gliese(\(starData?.value.gl_id ?? "")), BF(\(starData?.value.bayer_flamstedt ?? "")):")
